wip: add validation for int
